add user and his events deletion for admin fix common dialog finding for different users
